requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coverage==5.4
defusedxml==0.6.0
dj-rest-auth==2.1.3
Django==3.1.5
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
isort==5.7.0
kombu==4.6.11
oauthlib==3.1.0
phonenumbers==8.12.17
prompt-toolkit==3.0.15
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
